chore(TS): Create TS config for firebase
