+++ src/store/modules/easy-firestore/data-table-details/index.js
