+++ src/store/modules/easy-firestore/data-tables/index.js
