+++ src/store/modules/easy-firestore/get-gbq-to-gbq-confs/index.js
