+++ src/store/modules/easy-firestore/mirror-exc-gcs-to-gbq-conf-details/index.js
