+++ src/store/modules/easy-firestore/mirror-exc-gcs-to-gbq-confs/index.js
