+++ src/store/modules/easy-firestore/schemas/index.js
