+++ src/store/modules/easy-firestore/storage-to-storage-confs/index.js
