+++ src/store/modules/easy-firestore/storage-to-table-conf/index.js
