+++ src/store/modules/easy-firestore/storage-to-tables-confs/index.js
