+++ src/store/modules/easy-firestore/table-to-storage-confs/index.js
