+++ src/store/modules/easy-firestore/vm-launcher-confs/index.js
